バッチ処理にてDB書き込み確認。エラー処理、logging追加
